Make sure the auto build is triggered after a force kill
When an Xtext application is force killed, the Xtext will not have been
written (as it's written on shutdown). The applications thus forgets
the last built state during the next startup. The idea is that
the auto-builder will then pick up that the projects need to be rebuild.

However, for the latter we need to first touch the relevant projects.

Signed-off-by: Titouan Vervack <titouan.vervack@sigasi.com>